add standalone components for app.ts
