--- src/app/app-routing.module.ts
@@ -1,21 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-import { HomeComponent } from './home/home.component';
-import { AboutComponent } from './about/about.component';
-import { CoachesComponent } from './coaches/coaches.component';
-import { ProgramComponent } from './program/program.component';
-
-const routes: Routes = [
-  { path: '', component: HomeComponent },
-  { path: 'home', component: HomeComponent },
-  { path: 'about', component: AboutComponent },
-  { path: 'coaches', component: CoachesComponent },
-  { path: 'programs', component: ProgramComponent },
-  { path: '**', pathMatch: 'full', redirectTo: '/home' }
-];
-
-@NgModule({
-  imports: [RouterModule.forRoot(routes)],
-  exports: [RouterModule]
-})
-export class AppRoutingModule { }
